move files to `core` module
