+++ paperchat/core/__init__.py
@@ -1,0 +1,15 @@
+from .embeddings import EmbeddingModel, embed_doc_chunks
+from .ingestion import chunk_document, get_pdf_chunks, parse_pdf, process_document
+from .llm import RAGPipeline
+from .retrieval import similarity_search
+
+__all__ = [
+    "EmbeddingModel",
+    "RAGPipeline",
+    "chunk_document",
+    "embed_doc_chunks",
+    "get_pdf_chunks",
+    "parse_pdf",
+    "process_document",
+    "similarity_search",
+]
